Issue #157 feat : API 31 이상/미만 버전에서 스플래시 화면 표시
