+++ app/src/main/AndroidManifest.xml
@@ -14,27 +14,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.RightWeight.Splash"
         tools:targetApi="31">
-<!--        <activity-->
-<!--            android:name=".ui.TitleActivity"-->
-<!--            android:exported="true"-->
-<!--            android:screenOrientation="portrait"-->
-<!--            tools:ignore="LockedOrientationActivity">-->
-<!--            <intent-filter>-->
-<!--                <action android:name="android.intent.action.MAIN" />-->
-<!--                <category android:name="android.intent.category.LAUNCHER" />-->
-<!--            </intent-filter>-->
-
-<!--            <meta-data-->
-<!--                android:name="android.app.lib_name"-->
-<!--                android:value="" />-->
-<!--        </activity>-->
         <activity
             android:name=".ui.MainActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustPan"
+            android:theme="@style/Theme.RightWeight"
             tools:ignore="LockedOrientationActivity">
 
             <meta-data
@@ -46,6 +32,7 @@
             android:name=".ui.login.LoginActivity"
             android:screenOrientation="portrait"
             android:exported="true"
+            android:theme="@style/Theme.RightWeight.Splash"
             tools:ignore="LockedOrientationActivity">
 
             <intent-filter>
