+++ app/src/main/res/drawable/ic_splash_logo.xml
@@ -1,6 +1,6 @@
 <inset xmlns:android="http://schemas.android.com/apk/res/android"
     android:drawable="@drawable/img_right_weight"
-    android:insetLeft="72dp"
-    android:insetRight="72dp"
-    android:insetTop="90dp"
-    android:insetBottom="90dp"/>
+    android:insetLeft="18dp"
+    android:insetRight="18dp"
+    android:insetTop="36dp"
+    android:insetBottom="36dp"/>
